feat(v4): refactor OBS v4 Websocket
